Add rounding for getTotal() method
add rounding for total price to avoid payment conflicts because of longer floating numbers when handling net prices.

Example:
Backend uses net prices. When the net total is 10,86 € the total amount with taxes is calculated 12,9234 €. No further rounding seems to be applied. This caused an error for the Payone payment as the module tries to transmit: 'amount' => 1294.34